operaciones basicas

+++ 01-conceptosBasicos/04-problem1.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 int main()
 { 
-    int number1, number2;
+    float number1, number2;
 
     cout << "Ingrese un numero:" << endl;
     cin >> number1;
